+++ apps/aspiration/serializers.py
@@ -17,12 +17,13 @@ class CategorySerializer(serializers.ModelSerializer):
 
     def get_color(self, obj):
         mapping = {
-            'Infrastruktur': '#8e44ad',
-            'Keamanan': '#c0392b',
-            'Kebersihan': '#27ae60',
-            'Lainnya': '#7f8c8d',
+            'fasilitas': '#6D5DFF',
+            'lingkungan': '#9DC344',
+            'pendidikan': '#C03648',
+            'karakter': '#C08736',
+            'ibadah': '#BA36C0',
         }
-        return mapping.get(obj.name, '#3498db')
+        return mapping.get(obj.name.lower(), '#6C757D')
 
 
 class AspirationProgressSerializer(serializers.ModelSerializer):
@@ -53,6 +54,7 @@ class AspirationProgressSerializer(serializers.ModelSerializer):
 class AspirationSerializer(serializers.ModelSerializer):
     category_detail = CategorySerializer(source='category', read_only=True)
     progress_updates = AspirationProgressSerializer(many=True, read_only=True)
+    student_info = serializers.SerializerMethodField()
     
     category_id = serializers.PrimaryKeyRelatedField(
         queryset=Category.objects.all(),
@@ -91,12 +93,12 @@ class AspirationSerializer(serializers.ModelSerializer):
     
     status_display = serializers.CharField(source='get_status_display', read_only=True)
     status_color = serializers.SerializerMethodField()
-    sender = serializers.CharField(source='user.email', read_only=True)
+    student = serializers.CharField(source='user.email', read_only=True)
 
     class Meta:
         model = Aspiration
         fields = [
-            'id', 'report_id', 'sender', 'title', 'description', 'location',
+            'id', 'report_id', 'student', 'student_info', 'title', 'description', 'location',
             'category_id', 'category_detail', 
             'status', 'status_display', 'status_color',
             'image', 'video', 'progress_updates', 'created_at'
@@ -105,18 +107,53 @@ class AspirationSerializer(serializers.ModelSerializer):
 
     def get_status_color(self, obj):
         mapping = {
-            'menunggu': '#FFC107', 
-            'proses': '#007BFF',   
-            'selesai': '#28A745',  
-            'dibatalkan': '#DC3545'
+            'menunggu': '#EE9300', 
+            'proses': '#006EFF',   
+            'selesai': '#19A229',  
+            'dibatalkan': '#A21919'
         }
-        return mapping.get(obj.status, '#6C757D')
+        return mapping.get(obj.status.lower(), '#6C757D')
+    
+    def get_student_info(self, obj):
+        try:
+            student = obj.user.student_profile 
+            return {
+                "name": student.name,
+                "nis": student.nis,
+                "rombel": student.rombel,
+                "rayon": student.rayon,
+                # "image": student.image.url if student.image else None
+            }
+        except:
+            return None
 
 
 class NotificationSerializer(serializers.ModelSerializer):
     report_id = serializers.CharField(source='aspiration.report_id', read_only=True)
-
+    aspiration_title = serializers.CharField(source='aspiration.title', read_only=True)
+    aspiration_description = serializers.CharField(source='aspiration.description', read_only=True)
+    
+    student_info = serializers.SerializerMethodField()
+    student = serializers.CharField(source='aspiration.user.email', read_only=True)
+    
     class Meta:
         model = Notification
-        fields = ['id', 'report_id', 'message', 'is_read', 'created_at']
+        fields = [
+            'id',  'report_id', 'message', 'student', 'student_info',
+            'aspiration_title', 'aspiration_description', 
+            'is_read', 'created_at'
+        ]
         read_only_fields = ['id', 'report_id', 'message', 'created_at']
+
+    def get_student_info(self, obj):
+        try:
+            student_profile = obj.aspiration.user.student_profile 
+            return {
+                "name": student_profile.name,
+                "nis": student_profile.nis,
+                "rombel": student_profile.rombel,
+                "rayon": student_profile.rayon,
+                # "image": student.image.url if student.image else None
+            }
+        except:
+            return None