+++ config/urls.py
@@ -6,6 +6,7 @@ from django.conf.urls.static import static
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include(('apps.user.urls', 'users-api'))),
+    path('api/', include(('apps.aspiration.urls', 'aspiration-api'))),
 ]
 if settings.DEBUG:
     urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)